[Refactoring]struct별로 파일 분할
