+++ LadderGame/LadderGame/LadderPlayer.swift
@@ -1,0 +1,13 @@
+//
+//  LadderPlayer.swift
+//  LadderGame
+//
+//  Created by TaeHyeonLee on 2017. 10. 23..
+//  Copyright © 2017년 Codesquad Inc. All rights reserved.
+//
+
+import Foundation
+
+struct LadderPlayer {
+    let name : String
+}
